refactor(web): split teams filter controls
